python/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cmake==0.7.1
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.18.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
idna==2.8
numpy==1.16.3
Pillow==6.0.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyMySQL==0.9.3
pytesseract==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
tesseract==0.1.3
urllib3==1.25.3
